+++ src/app/(tests)/freetest/FreeTestContainer.tsx
@@ -4,66 +4,10 @@ import { useState } from 'react';
 import './_FreeTestPage.scss';
 import FileUpLoader from '@/components/FileUpload/FileUploader';
 import ProtectedRoute from '@/components/ProtectedRoute/ProtectedRoute';
+import { shareURL } from '@/utils/shareURL';
 
 function FreeTestContainer({ totalCount }: { totalCount: number }) {
   const [currentStep, setCurrentStep] = useState(1);
-
-  // 공유 함수 예외처리
-  const handleShareURL = () => {
-    const text = `[ 하이라이톡 - 우리는 어떻게 대화하고 있을까? ]\n🤖하이라이톡 에서 테스트 해보세요!\n\nURL - ${location.href}`;
-
-    //모바일
-    if (navigator.share) {
-      navigator
-        .share({
-          title: '하이라이톡 - AI 분석 결과',
-          text: text,
-          url: location.href,
-        })
-        .then(() => {
-          console.log('공유 성공!');
-        })
-        .catch((error) => {
-          console.error('공유 실패:', error);
-        });
-    } else if (navigator.clipboard) {
-      navigator.clipboard
-        .writeText(text)
-        .then(() => {
-          showToast('클립보드에 URL이 복사되었어요!');
-        })
-        .catch((error) => {
-          alert('클립보드에 복사하는데 실패했어요. 직접 복사해 주세요: ' + text);
-        });
-    }
-  };
-
-  // 토스트 알림 함수
-  const showToast = (message: string) => {
-    const toast = document.createElement('div');
-    toast.textContent = message;
-    toast.style.position = 'fixed';
-    toast.style.bottom = '20px';
-    toast.style.left = '50%';
-    toast.style.transform = 'translateX(-50%)';
-    toast.style.backgroundColor = 'rgba(0, 0, 0, 0.8)';
-    toast.style.color = 'white';
-    toast.style.padding = '10px 20px';
-    toast.style.borderRadius = '5px';
-    toast.style.zIndex = '1000';
-    toast.style.opacity = '1';
-    toast.style.transition = 'opacity 0.5s ease-in-out';
-
-    document.body.appendChild(toast);
-
-    // 3초 후에 토스트 메시지 사라짐
-    setTimeout(() => {
-      toast.style.opacity = '0';
-      setTimeout(() => {
-        document.body.removeChild(toast);
-      }, 500); // 0.5초 후에 DOM에서 제거
-    }, 3000); // 3초 동안 토스트 표시
-  };
 
   return (
     <div className="test-page-cover">
@@ -77,7 +21,12 @@ function FreeTestContainer({ totalCount }: { totalCount: number }) {
               </h2>
               <p className="sub-p">하이라이톡에서 대화내용을 업로드 해 AI분석을 시작 하세요!</p>
               <div className="test-page-action">
-                <button onClick={() => handleShareURL()} className="btn n1">
+                <button
+                  onClick={() => {
+                    shareURL('🤖하이라이톡에서 테스트 해보세요!');
+                  }}
+                  className="btn n1"
+                >
                   공유하기
                 </button>
                 <button onClick={() => setCurrentStep(2)} className="btn n2">
